[EKA-000] Revert handling registration of EventBus in BaseActivity, change impl of RedirectInterceptor starting activity
