--- commons/utils/src/main/java/in/projecteka/jataayu/util/event/UserUnauthorizedRedirectEvent.kt
@@ -1,5 +1,0 @@
-package `in`.projecteka.jataayu.util.event
-
-enum class UserUnauthorizedRedirectEvent {
-    REDIRECT
-}
